[#5] 리뷰반영: 로그인 관련 서비스 추상화
- 로그인 방식 변경 및 확장을 고려해 인터페이스 레이어 추가
- BasicLoginService(세션없는 기본 로그인 방식) 구현
- 추후 Redis 활용한 로그인 서비스 구현 예정
